SCH-55 refactored google connection process to be more modular
