+++ backend/src/models/calendar.go
@@ -1,9 +1,9 @@
 package models
 
 // Define a struct that represents a Google calendar
-type CalendarEntry struct {
+type Calendar struct {
 	Id       string `json:"id"`
-	Location     string `json:"location"`
+	Location string `json:"location"`
 	Summary  string `json:"summary"`
 	TimeZone string `json:"timeZone"`
 }
