+++ backend/src/routes/routes.go
@@ -2,6 +2,7 @@ package routes
 
 import (
 	handlers "schedulii/src/handlers"
+	google "schedulii/src/handlers/google"
 	"schedulii/src/middleware"
 
 	"github.com/gin-gonic/contrib/static"
@@ -13,20 +14,17 @@ func SetupRoutes(engine *gin.Engine) {
 	engine.Use(static.Serve("/", static.LocalFile("../../frontend/build", true)))
 
 	engine.GET("/health", handlers.HealthCheck)
-	engine.GET("/events", handlers.GetUserCalendarEvents)
-	engine.GET("/calendars", handlers.GetCalendarList)
 
 	googleAuth := engine.Group("/google")
 	{
-		googleAuth.GET("/googleAuth", handlers.RunGoogleConnection)
-		googleAuth.GET("/googleCallback", handlers.RunGoogleCallback)
+		googleAuth.GET("/googleAuth", google.RunGoogleConnection)
+		googleAuth.GET("/googleCallback", google.RunGoogleCallback)
+		googleAuth.GET("/calendars", google.UserCalendarListHandler)
+		googleAuth.GET("/events", google.UserCalendarEventsHandler)
 	}
 
 	authorized := engine.Group("/authorized")
 	authorized.Use(middleware.CheckAuthenticated)
-	{
-		authorized.GET("", handlers.GetCalendars)
-	}
 
 	login := engine.Group("/login")
 	{
